requirements.txt
requests
bs4
feedparser
arrow
pymongo
unidecode
python-dotenv
fastapi
uvicorn
mangum